feat(package): add common save function to lib
